requirements.txt
opencc-python-reimplemented
csv
